# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Vita API doc",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/admin/approve/{id}": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Altera o status do enfermeiro para 'verificado' (verification_seal: true). Requer autenticação de Admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Aprova o cadastro de um enfermeiro",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Enfermeiro a ser aprovado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enfermeiro aprovado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponseString"
                        }
                    },
                    "400": {
                        "description": "ID inválido ou erro na aprovação",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Não autorizado (Token JWT inválido ou ausente)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Proibido (Usuário não é Administrador)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/dashboard": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna as principais métricas e KPIs para a tela de dashboard do administrador. Requer autenticação de Admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Dados do Dashboard Administrativo",
                "responses": {
                    "200": {
                        "description": "Dados de dashboard carregados com sucesso",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessDashboardResponse"
                        }
                    },
                    "400": {
                        "description": "Erro ao carregar dados",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Não autorizado (Token JWT inválido ou ausente)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Proibido (Usuário não é Administrador)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/documents/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna uma lista de documentos (com URLs de download) de um enfermeiro específico para aprovação. Requer autenticação de Admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Obtém documentos do enfermeiro para análise",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Enfermeiro (Nurse ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Documentos retornados com sucesso",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessDocumentsResponse"
                        }
                    },
                    "400": {
                        "description": "ID inválido ou erro ao buscar documentos",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Não autorizado (Token JWT inválido ou ausente)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Proibido (Usuário não é Administrador)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/download/{id}": {
            "get": {
                "description": "Faz o download de um arquivo (como um documento de enfermeiro) com base no seu ObjectID do GridFS.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Baixa um arquivo do GridFS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Arquivo (File ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "O arquivo para download",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "ID inválido ou Arquivo não encontrado",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro ao enviar o arquivo",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/reject/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Envia um email ao enfermeiro com o motivo da rejeição. Requer autenticação de Admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Rejeita o cadastro de um enfermeiro",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Enfermeiro a ser rejeitado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Objeto contendo a descrição/motivo da rejeição",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RejectDescription"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enfermeiro rejeitado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponseString"
                        }
                    },
                    "400": {
                        "description": "ID inválido, corpo da requisição inválido ou erro na rejeição",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Não autorizado (Token JWT inválido ou ausente)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Proibido (Usuário não é Administrador)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/user/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite ao administrador deletar um usuário (Paciente ou Enfermeiro) permanentemente. Requer autenticação de Admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Deleta um usuário (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Usuário a ser deletado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário deletado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponseNoData"
                        }
                    },
                    "400": {
                        "description": "ID inválido ou erro ao deletar",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Não autorizado (Token JWT inválido ou ausente)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Proibido (Usuário não é Administrador)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite ao administrador atualizar campos de um usuário (Paciente ou Enfermeiro). Requer autenticação de Admin.\nCampos protegidos (id, created_at, updated_at) não podem ser atualizados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Atualiza um usuário (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Usuário a ser atualizado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos para atualizar (JSON arbitrário). Ex: {\\",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessUserTypeResponse"
                        }
                    },
                    "400": {
                        "description": "JSON inválido, ID não encontrado ou campo protegido",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Não autorizado (Token JWT inválido ou ausente)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Proibido (Usuário não é Administrador)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna listas completas de usuários, enfermeiros e visitas para gerenciamento do admin. Requer autenticação de Admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Listas de gerenciamento (Usuários, Enfermeiros, Visitas)",
                "responses": {
                    "200": {
                        "description": "Listas retornadas com sucesso",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessUserListsResponse"
                        }
                    },
                    "400": {
                        "description": "Erro ao buscar listas",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Não autorizado (Token JWT inválido ou ausente)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Proibido (Usuário não é Administrador)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.DocumentInfoResponse": {
            "type": "object",
            "properties": {
                "download_url": {
                    "description": "A URL para baixar o arquivo",
                    "type": "string"
                },
                "image_id": {
                    "type": "string"
                },
                "name": {
                    "description": "Um nome amigável, ex: \"Documento de Licença (COREN)\"",
                    "type": "string"
                },
                "type": {
                    "description": "Um identificador, ex: \"license_document\"",
                    "type": "string"
                }
            }
        },
...
        "dto.RejectDescription": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
...
        "dto.UserTypeResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cpf": {
                    "description": "PixKey      string `json:\"pix_key\"`",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_access": {
                    "type": "boolean"
                },
                "hidden": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Descrição do erro"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "utils.SuccessDashboardResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.DashboardAdminDataResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Dados de dashboard carregados com sucesso"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "utils.SuccessDocumentsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Note que é um slice []",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DocumentInfoResponse"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Documentos retornados com sucesso"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "utils.SuccessResponseNoData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operação realizada com sucesso"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "utils.SuccessResponseString": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "Mensagem de dados"
                },
                "message": {
                    "type": "string",
                    "example": "Operação realizada com sucesso"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "utils.SuccessUserListsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.UserListsResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Listas de usuários retornadas com sucesso."
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "utils.SuccessUserTypeResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.UserTypeResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Usuário atualizado com sucesso."
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "\"Para autorizar, insira seu token JWT precedido de 'Bearer '. Exemplo: Bearer \u003cseu-token-aqui\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}